refactor : entity 패키지 profile.entity로 이동
