+++ src/main/java/com/meoguri/linkocean/domain/notification/service/NotificationServiceImpl.java
@@ -15,8 +15,8 @@ import com.meoguri.linkocean.domain.notification.entity.Notification;
 import com.meoguri.linkocean.domain.notification.entity.vo.NotificationType;
 import com.meoguri.linkocean.domain.notification.persistence.NotificationRepository;
 import com.meoguri.linkocean.domain.notification.service.dto.ShareNotificationCommand;
-import com.meoguri.linkocean.domain.profile.command.entity.Profile;
 import com.meoguri.linkocean.domain.profile.command.persistence.FindProfileByIdQuery;
+import com.meoguri.linkocean.domain.profile.entity.Profile;
 
 import lombok.RequiredArgsConstructor;
