+++ src/main/java/com/meoguri/linkocean/domain/profile/command/persistence/ProfileRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import com.meoguri.linkocean.domain.profile.command.entity.Profile;
+import com.meoguri.linkocean.domain.profile.entity.Profile;
 
 public interface ProfileRepository extends JpaRepository<Profile, Long>, CustomProfileRepository {
