+++ src/main/java/com/meoguri/linkocean/domain/profile/query/persistence/ProfileDao.java
@@ -4,7 +4,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.repository.Repository;
 
-import com.meoguri.linkocean.domain.profile.command.entity.Profile;
+import com.meoguri.linkocean.domain.profile.entity.Profile;
 import com.meoguri.linkocean.domain.profile.query.persistence.dto.ProfileFindCond;
 
 public interface ProfileDao extends Repository<Profile, Long> {
