+++ src/main/java/com/meoguri/linkocean/domain/user/service/UserService.java
@@ -1,6 +1,6 @@
 package com.meoguri.linkocean.domain.user.service;
 
-import com.meoguri.linkocean.domain.profile.command.entity.Profile;
+import com.meoguri.linkocean.domain.profile.entity.Profile;
 import com.meoguri.linkocean.domain.user.entity.vo.Email;
 import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
 import com.meoguri.linkocean.domain.user.service.dto.GetUserResult;
