+++ src/test/java/com/meoguri/linkocean/domain/bookmark/entity/BookmarkTest.java
@@ -2,7 +2,7 @@ package com.meoguri.linkocean.domain.bookmark.entity;
 
 import static com.meoguri.linkocean.domain.bookmark.entity.vo.Category.*;
 import static com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType.*;
-import static com.meoguri.linkocean.domain.profile.command.entity.Profile.*;
+import static com.meoguri.linkocean.domain.profile.entity.Profile.*;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
@@ -16,7 +16,7 @@ import org.springframework.test.util.ReflectionTestUtils;
 import com.meoguri.linkocean.domain.bookmark.entity.vo.Category;
 import com.meoguri.linkocean.domain.bookmark.entity.vo.OpenType;
 import com.meoguri.linkocean.domain.bookmark.entity.vo.TagIds;
-import com.meoguri.linkocean.domain.profile.command.entity.Profile;
+import com.meoguri.linkocean.domain.profile.entity.Profile;
 import com.meoguri.linkocean.domain.tag.entity.Tag;
 import com.meoguri.linkocean.test.support.domain.entity.BaseEntityTest;
