+++ src/test/java/com/meoguri/linkocean/domain/profile/command/persistence/FindProfileByIdQueryTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Import;
 
-import com.meoguri.linkocean.domain.profile.command.entity.Profile;
+import com.meoguri.linkocean.domain.profile.entity.Profile;
 import com.meoguri.linkocean.test.support.domain.persistence.BasePersistenceTest;
 
 @Import(FindProfileByIdQuery.class)
